// app/requests/_personal/getModerationClient/getModerationClientSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetModerationClientSuccess {
  data: {
    id: number;
    phone: number;
    email: string;
    logo: null;
    project: {
      id: number;
      name: string;
      brand: unknown[];
    }[];
    place: {
      id: number;
      name: string;
      latitude: string;
      longitude: string;
      address_kladr: string;
    }[];
    roles: {
      id: number;
      name: "client" | "manager" | "recruiter" | "specialist" | "supervisor";
    }[];
  }[];
  links: {
    first: string;
    last: null;
    prev: string;
    next: null;
  };
  meta: {
    current_page: number;
    from: number | null;
    path: string;
    per_page: number;
    to: number | null;
  };
}